moving interpolators and adding fast interpolator posibility
