+++ spyro/solvers/wave.py
@@ -338,6 +338,7 @@ class Wave(Model_parameters, metaclass=ABCMeta):
         new_file=None,
         output=False,
         dg_velocity_model=True,
+        fast_interpolate=False,
     ):
         """Method to define new user velocity model or file. It is optional.
 
@@ -388,7 +389,7 @@ class Wave(Model_parameters, metaclass=ABCMeta):
             self.initial_velocity_model = velocity_model_function
         elif new_file is not None:
             self.initial_velocity_model_file = new_file
-            self._initialize_model_parameters()  # TODO in PR206
+            self._initialize_model_parameters(fast_interpolate=fast_interpolate)  # TODO in PR206
         elif constant is not None:
             V = self.function_space
             vp = fire.Function(V, name="velocity")
